Fix 'logout'
I've also changed the type of the 'Permissions' field.

'int' is enough to store all of our permissions!

+++ go/main.go
@@ -11,7 +11,7 @@ import (
 )
 
 type Session struct {
-	Permissions uint8
+	Permissions int
 }
 
 var cookies = true
